+++ NEWS.md
@@ -22,6 +22,8 @@ Following functions are now defunct:
 
 ## Changes to functions
 
+* `plot_model()` gets a `show.zeroinf`-argument to show or hide the zero-inflation-part of models in the plot.
+* `plot_model()` now plots the zero-inflated part of the model for `brmsfit`-objects.
 * The `grid.breaks`-argument for `plot_model()` and `plot_models()` now also takes a vector of values to directly define the grid breaks for the plot.
 * Better default calculation for grid breaks in `plot_model()` and `plot_models()` when the `grid.breaks`-argument is of length one.
 * The `terms`-argument for `plot_model()` now also allows the specification of a range of numeric values in square brackets for marginal effects plots, e.g. `terms = "age [30:50]"`.